requirements/common.txt
pandas~=1.5.2
openpyxl~=3.0.10
flask~=2.2.2
flask-sqlalchemy~=3.0.2
flask-migrate~=4.0.0